move data module to domain module
