+++ app/build.gradle.kts
@@ -47,7 +47,6 @@ dependencies {
     implementation(fileTree(mapOf("dir" to "libs", "include" to listOf("*.jar"))))
 
     implementation(project(":core"))
-    implementation(project(":libraries:data"))
     implementation(project(":libraries:domain"))
     implementation(project(":presentation"))
     implementation(project(":shopping_list"))
