+++ core/src/main/java/com/zizohanto/android/tobuy/core/di/DataModule.kt
@@ -1,9 +1,9 @@
 package com.zizohanto.android.tobuy.core.di
 
-import com.zizohanto.android.tobuy.data.repository.ProductRepositoryImpl
-import com.zizohanto.android.tobuy.data.repository.ShoppingListRepositoryImpl
 import com.zizohanto.android.tobuy.domain.repository.ProductRepository
+import com.zizohanto.android.tobuy.domain.repository.ProductRepositoryImpl
 import com.zizohanto.android.tobuy.domain.repository.ShoppingListRepository
+import com.zizohanto.android.tobuy.domain.repository.ShoppingListRepositoryImpl
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
