+++ libraries/domain/src/main/AndroidManifest.xml
