+++ libraries/domain/src/main/java/com/zizohanto/android/tobuy/domain/mappers/ProductEntityMapper.kt
@@ -1,8 +1,8 @@
-package com.zizohanto.android.tobuy.data.mappers
+package com.zizohanto.android.tobuy.domain.mappers
 
-import com.zizohanto.android.tobuy.data.mappers.base.EntityMapper
-import com.zizohanto.android.tobuy.data.models.ProductEntity
+import com.zizohanto.android.tobuy.domain.mappers.base.EntityMapper
 import com.zizohanto.android.tobuy.domain.models.Product
+import com.zizohanto.android.tobuy.domain.models.ProductEntity
 import javax.inject.Inject
 
 class ProductEntityMapper @Inject constructor() : EntityMapper<ProductEntity, Product> {
