+++ libraries/domain/src/main/java/com/zizohanto/android/tobuy/domain/models/ProductEntity.kt
@@ -1,4 +1,4 @@
-package com.zizohanto.android.tobuy.data.models
+package com.zizohanto.android.tobuy.domain.models
 
 import java.util.UUID
